update version: 1.70.1

+++ ClashX/General/Managers/ClashResourceManager.swift
@@ -2,6 +2,7 @@
 import Alamofire
 import AppKit
 import Foundation
+import Gzip
 
 class ClashResourceManager {
     static func check() -> Bool {
@@ -31,14 +32,20 @@ class ClashResourceManager {
         if fileManage.fileExists(atPath: destMMDBPath) {
             let vaild = verifyGEOIPDataBase().toBool()
             let versionChange = AppVersionUtil.hasVersionChanged || AppVersionUtil.isFirstLaunch
-            if !vaild || versionChange {
+            let customMMDBSet = !Settings.mmdbDownloadUrl.isEmpty
+            if !vaild || (versionChange && customMMDBSet) {
                 try? fileManage.removeItem(atPath: destMMDBPath)
             }
         }
 
         if !fileManage.fileExists(atPath: destMMDBPath) {
-            if let mmdbPath = Bundle.main.path(forResource: "Country", ofType: "mmdb") {
-                try? fileManage.copyItem(at: URL(fileURLWithPath: mmdbPath), to: URL(fileURLWithPath: destMMDBPath))
+            if let mmdbUrl = Bundle.main.url(forResource: "Country.mmdb", withExtension: "gz") {
+                do {
+                    let data = try Data(contentsOf: mmdbUrl).gunzipped()
+                    try data.write(to: URL(fileURLWithPath: destMMDBPath))
+                } catch let err {
+                    Logger.log("add mmdb fail:\(err)", level: .error)
+                }
             }
         }
     }
@@ -61,21 +68,47 @@ extension ClashResourceManager {
     }
 
     @objc private static func updateGeoIP() {
-        let url = "https://static.clash.to/GeoIP2/GeoIP2-Country.mmdb"
-        AF.download(url) { (_, _) -> (destinationURL: URL, options: DownloadRequest.Options) in
+        guard let url = showCustomAlert() else { return }
+        AF.download(url, to:  { (_, _) in
             let path = kConfigFolderPath.appending("/Country.mmdb")
             return (URL(fileURLWithPath: path), .removePreviousFile)
-        }.response { res in
-            let title = NSLocalizedString("Update GEOIP Database", comment: "")
-            let info: String
+        }).response { res in
+            var info: String
             switch res.result {
             case .success:
                 info = NSLocalizedString("Success!", comment: "")
+                Logger.log("update success")
             case let .failure(err):
                 info = NSLocalizedString("Fail:", comment: "") + err.localizedDescription
+                Logger.log("update fail \(err)")
             }
-            NSUserNotificationCenter.default.post(title: title, info: info)
-            checkMMDB()
+            if !verifyGEOIPDataBase().toBool() {
+                info = "Database verify fail"
+                checkMMDB()
+            }
+            let alert = NSAlert()
+            alert.messageText = NSLocalizedString("Update GEOIP Database", comment: "")
+            alert.informativeText = info
+            alert.runModal()
         }
+    }
+    
+    private static func showCustomAlert() -> String? {
+        let alert = NSAlert()
+        alert.messageText = NSLocalizedString("Custom your GEOIP MMDB download address.", comment: "")
+        let inputView = NSTextField(frame: NSRect(x: 0, y: 0, width: 250, height: 24))
+        inputView.placeholderString =  "https://github.com/Dreamacro/maxmind-geoip/releases/latest/download/Country.mmdb"
+        inputView.stringValue = Settings.mmdbDownloadUrl
+        alert.accessoryView = inputView
+        alert.addButton(withTitle: NSLocalizedString("OK", comment: ""))
+        alert.addButton(withTitle: NSLocalizedString("Cancel", comment: ""))
+        if alert.runModal() == .alertFirstButtonReturn {
+            if inputView.stringValue.isEmpty {
+                return inputView.placeholderString
+            }
+            Settings.mmdbDownloadUrl = inputView.stringValue
+            return inputView.stringValue
+        }
+        return nil
     }
 }
